+++ util/jsonapi/preimages.go
@@ -10,12 +10,14 @@ import (
 	"io"
 
 	"github.com/ethereum/go-ethereum/common"
-
-	"github.com/offchainlabs/nitro/validator"
 )
 
 type PreimagesMapJson struct {
 	Map map[common.Hash][]byte
+}
+
+func NewPreimagesMapJson(inner map[common.Hash][]byte) PreimagesMapJson {
+	return PreimagesMapJson{inner}
 }
 
 func (m *PreimagesMapJson) MarshalJSON() ([]byte, error) {
@@ -167,88 +169,4 @@ func (m *PreimagesMapJson) UnmarshalJSON(data []byte) error {
 		}
 	}
 	return nil
-}
-
-type BatchInfoJson struct {
-	Number  uint64
-	DataB64 string
-}
-
-type ValidationInputJson struct {
-	Id            uint64
-	HasDelayedMsg bool
-	DelayedMsgNr  uint64
-	PreimagesB64  PreimagesMapJson
-	BatchInfo     []BatchInfoJson
-	DelayedMsgB64 string
-	StartState    validator.GoGlobalState
-}
-
-func ValidationInputToJson(entry *validator.ValidationInput) *ValidationInputJson {
-	res := &ValidationInputJson{
-		Id:            entry.Id,
-		HasDelayedMsg: entry.HasDelayedMsg,
-		DelayedMsgNr:  entry.DelayedMsgNr,
-		DelayedMsgB64: base64.StdEncoding.EncodeToString(entry.DelayedMsg),
-		StartState:    entry.StartState,
-		PreimagesB64:  PreimagesMapJson{entry.Preimages},
-	}
-	for _, binfo := range entry.BatchInfo {
-		encData := base64.StdEncoding.EncodeToString(binfo.Data)
-		res.BatchInfo = append(res.BatchInfo, BatchInfoJson{binfo.Number, encData})
-	}
-	return res
-}
-
-func ValidationInputFromJson(entry *ValidationInputJson) (*validator.ValidationInput, error) {
-	valInput := &validator.ValidationInput{
-		Id:            entry.Id,
-		HasDelayedMsg: entry.HasDelayedMsg,
-		DelayedMsgNr:  entry.DelayedMsgNr,
-		StartState:    entry.StartState,
-		Preimages:     entry.PreimagesB64.Map,
-	}
-	delayed, err := base64.StdEncoding.DecodeString(entry.DelayedMsgB64)
-	if err != nil {
-		return nil, err
-	}
-	valInput.DelayedMsg = delayed
-	for _, binfo := range entry.BatchInfo {
-		data, err := base64.StdEncoding.DecodeString(binfo.DataB64)
-		if err != nil {
-			return nil, err
-		}
-		decInfo := validator.BatchInfo{
-			Number: binfo.Number,
-			Data:   data,
-		}
-		valInput.BatchInfo = append(valInput.BatchInfo, decInfo)
-	}
-	return valInput, nil
-}
-
-type MachineStepResultJson struct {
-	Hash        common.Hash
-	Position    uint64
-	Status      uint8
-	GlobalState validator.GoGlobalState
-}
-
-func MachineStepResultToJson(result *validator.MachineStepResult) *MachineStepResultJson {
-	return &MachineStepResultJson{
-		Hash:        result.Hash,
-		Position:    result.Position,
-		Status:      uint8(result.Status),
-		GlobalState: result.GlobalState,
-	}
-}
-
-func MachineStepResultFromJson(resultJson *MachineStepResultJson) (*validator.MachineStepResult, error) {
-
-	return &validator.MachineStepResult{
-		Hash:        resultJson.Hash,
-		Position:    resultJson.Position,
-		Status:      validator.MachineStatus(resultJson.Status),
-		GlobalState: resultJson.GlobalState,
-	}, nil
 }
